--- a/code/Julia/StrFileRead.jl
+++ b/code/Julia/StrFileRead.jl
@@ -2,13 +2,6 @@
 # Read Edges from file
 #
 function StrFileRead(fname)
-  fid = open(fname,"r");
-  uv = split(readall(fid));
-  # ut = int(uv[1:2:end])';   # makt it horizontal vector
-  # vt = int(uv[2:2:end])';   # makt it horizontal vector
-  ut = [parse(Int64,s) for s=uv[1:2:end]] # makt it horizontal vector
-  vt = [parse(Int64,s) for s=uv[2:2:end]] # makt it horizontal vector
-  close(fid)
-  return ut,vt
+    uv = readdlm(fname, Int)
+    return uv[1:2:end], uv[2:2:end]
 end
-