--- a/code/Julia/StrFileRead.jl
+++ b/code/Julia/StrFileRead.jl
@@ -2,6 +2,7 @@
 # Read Edges from file
 #
 function StrFileRead(fname)
-    uv = readdlm(fname, Int)
-    return uv[1:2:end], uv[2:2:end]
+    uv = readdlm(fname, ' ', Int)::Matrix{Int}
+    n = size(uv, 2)
+    return uv[1:2:n], uv[2:2:n]
 end