The Julia code in the repository is not very idiomatic and contains a lot of inefficient MATLAB-like commands.

This pull request makes several code changes to improve the speed of the Julia code. In many cases the improvement is 10x or greater.

Summary of changes:

- In-place mutating operations are used where possible, e.g. `scale!(a, b)` instead of `a = a.*b`, and array views `sub(A, :, 1)` in lieu of copies `A[:, 1]`.
- Simplified file I/O code
- Avoid vectorized operations in favor of writing out explicit for loops
- Use `scale!` instead of sparse matmul for diagonal x sparse in K2
- Remove type instability - don't reuse `i` for scalar and vector indexes, and so on

I also removed semicolons in the lines that I touched - they are mostly unnecessary in Julia and they are not needed to suppress output or to end a statement.

The one thing I didn't change is to rewrite `b*A` into `(A*b)'`, since it looks like you are explicitly benchmarking "vecmat" operations, which are not natural in Julia and have to be simulated with row matrix x matrix operations.

after:
```
scale	k0-edges-per-sec	k1-edges-per-sec	k2-edges-per-sec	k3-edges-per-sec
16384	1083001.8028436995	29322.508330906006	1805017.561022902	3871300.523289533
32768	1773909.8612562993	990937.6142354382	1996761.7214538117	362220034.74296784
65536	1859974.6932642444	1146452.2013610622	2055791.2561303037	250749012.0463673
131072	1717525.4529797793	744660.0460030059	1462645.891735659	180865515.28740132
262144	1413516.7398048237	745206.9510094854	1340359.7984361553	243721324.43513978
524288	1533899.1827641048	746738.5727661067	1219988.3415962423	216147601.31171572
1048576	1591695.317053853	764024.3808543442	1171522.541363742	143490752.44497326
```

before:
```
scale	k0-edges-per-sec	k1-edges-per-sec	k2-edges-per-sec	k3-edges-per-sec
16384	32715.994262927718	49560.807274887644	49775.781129473966	1756696.9662404153
32768	196089.1794055616	175006.01276075182	1184763.695901588	103459555.1693784
65536	194237.61984653943	213860.2440404617	1002932.0642994292	81663683.21827883
131072	217681.09766689962	222660.69418494092	989774.7341211278	54229594.55421871
262144	218191.06603552564	221851.16735839212	581347.1172400148	59082335.998872004
524288	206518.05565072008	221310.01817073236	986361.732802518	91955407.56957653
1048576	218349.76349001713	235138.21278212997	987605.8729298214	91022591.21137235
```